Make configuration of Spring IO plugin less invasive
The plugin is now only applied when the build is run with the
platformVersion property. This property is also used to determine
the version of the platform-versions dependency that will be used
by the springIoCheck task. This breaks a circular dependency between
the projects that are included in the Platform and the Platform
itself.